finalisation des composants d'achat sur UVAR

+++ database/migrations/2021_01_16_083531_create_transactions_table.php
@@ -15,7 +15,7 @@ class CreateTransactionsTable extends Migration
     {
         Schema::create('transactions', function (Blueprint $table) {
             $table->bigIncrements('id');
-            $table->float('amount');
+            $table->decimal('amount', 12, 2);
             $table->string('operator');
             $table->string('operator_id_transaction');
             $table->string('description')->nullable();
